refactor(driver): extract rewind methods to driver/rewind.rs (refactor 2.1g)
Second impl-block extraction of Phase 2.1. Moves the four rewind
methods out of `impl Driver` in mod.rs into a sibling module.

- `crates/based-rollup/src/driver/rewind.rs` (new, 201 lines):
  - Full `impl<P, Pool> Driver<P, Pool> where …` block with the same
    where-clause as mod.rs.
  - Methods migrated:
    - `clear_internal_state` — drop preconfirmed_hashes,
      pending_submissions, pending_l1, hold, and forward-tx queue.
    - `set_rewind_target` — record earliest rewind target for a batch.
    - `rewind_to_re_derive` — canonical 8-step hard rewind sequence
      that owns invariants #9 and #10.
    - `rewind_l2_chain` — async FCU-based unwind of the reth L2 tip,
      including the "reth refuses to unwind committed blocks" fallback
      that marks them immutable.
  - All methods are `pub(super) fn` so the rest of the `impl Driver`
    tree (verify/flush/step_builder/etc.) can call them via `self`.
  - Imports: `super::Driver`,
    `super::types::{DriverMode, FORK_CHOICE_DEPTH, compute_forkchoice_state}`,
    `eyre::{Result, WrapErr}`, reth provider bounds,
    `std::collections::VecDeque`, `tracing::{info, warn}`.

- `crates/based-rollup/src/driver/mod.rs` (-143 lines → 4244 total):
  - `mod rewind;` declaration.
  - Four method bodies deleted from the main `impl Driver` block.
  - Obsolete "Save the L1-confirmed anchor to the L2 database."
    doc comment removed (it was dangling after the earlier 2.1f
    journal extraction removed the method it described).

Cumulative progress:
- mod.rs: 5509 → 4244 lines (-1265 / -23.0%)
- 6 sibling modules totaling 1452 lines.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Next: 2.1h extracts `build_and_insert_block`, `fork_choice_updated_with_retry`,
and `update_fork_choice` to `driver/build.rs`. These methods are the
block-insertion + engine-API layer used by every step_* function.
